Have a go at fixing class MotorsBus(Protocol)
In the Robot class, we have the comment:

```python
    # TODO(rcadene, aliberts): Add unit test checking the protocol is implemented in the corresponding classes
```

This is a proof of concept that it is possible to get a type checker to check
our Protocol implementations.

In my case, I was using the pylance integration in vscode, and ignoring all of the other
warnings and errors. I can have a crack at adding a mypy check in CI, but it might be quite
invasive, and I haven't thought about how to split the work up and do it incrementally.

Making MotorsBus be an abstract base class would be an alternative way to do this.
That would have the advantage of making "go to implementations" work in IDEs
(I've not tested that approach yet). The disadvantage of abstract base class is
that it becomes tempting to start using the Template Method Pattern and move
some logic into the abstract base class (this can get difficult to reason about
**really** quickly). If you want me to have a go at that for compariasson, I can.

If you like this static type checking approach, I can try cleaning this PR up.
Things that would need to be done include:

- [ ] clean up the `values` type and return type for apply_calibration() and revert_calibration()
  - [ ] maybe make a helper type to avoid repeating `NDArray[np.int64] | NDArray[np.float64] | list[int] | list[float]` everywhere
- [ ] make mypy (or some other tool) actually check the relevant files, and make it easy to gradually grow the mypy coverage to more of the codebase

A follow-up PR could then do:
- [ ] apply the pattern to the other places that implement the MotorsBus protocol implicitly
- [ ] also apply the pattern to the Robot protocol