Path to card_template.md passed in relation to the current location of the lerobot package.
In the previous version the path was served statically and relative to the current folder: `./lerobot/common/datasets/card_template.md`
This creates problems when launching LerobotDataset.push_to_hub() outside of the package folder.

In the current version the path is provided in relation to the current path of the package: `importlib.resources.path("lerobot.common.datasets", "card_template.md")`
This allows to execute the method `create_lerobot_dataset_card` from any folder. As long as Lerobot is installed.

On branch fix--dataset_push_to_hub
Changes to be committed:
	modified:   lerobot/common/datasets/utils.py