feat: add support for external plugin config dataclasses
The existing `ChoiceRegistry`-based config classes are limited by
requiring that all subclasses be imported in order to be registered.
This means that in order to use the LeRobot training scripts, any
external library that defines its own configs MUST be imported,
which requires modifications of this repos files.

In order to make the training scripts usable without modification,
this commit introduces CLI argument parsing to load external plugins
at runtime. This will import plugins and trigger registration of
externally defined config dataclasses with the base choice registry.