feat: add remove_episodes utility
This commit introduces a remove_episodes function/CLI tool to remove
specific episodes from a dataset, and will automatically modify
all required data, video, and metadata.

The function will safely remove the episodes, meaning that if at
any point during the process a failure occurs, the original dataset
is preserved. Additionally, the original dataset is optionally backed up
in case it is needed to revert to.