Adding support for so100 and so101 URDFs in rerun
The main change in this commit, besides adding a URDF directory and files,
is the URDFLogger class. This class provides the methods to log the URDF
to Rerun as a static asset. It will also log the joint angles. The robot
type is needed to resolve the motor norm modes, which default to M100_100
if the robot type does not have a motor bus.